BREAKING CHANGE: split and rename metricq.types
Separate modules are easier to maintain.
The "types" name caused problems with mypy.

If you consider "metricq.types" to have been public interface, as we advertise in the examples, then this is a breaking change.

Fixes #154
