+++ metricq/history_client.py
@@ -47,7 +47,7 @@ from .exceptions import (
 )
 from .logging import get_logger
 from .rpc import rpc_handler
-from .types import TimeAggregate, Timedelta, Timestamp, TimeValue
+from .timeseries import TimeAggregate, Timedelta, Timestamp, TimeValue
 from .version import __version__  # noqa: F401 - shut up flake8, automatic version str
 
 logger = get_logger(__name__)
