+++ metricq/source.py
@@ -41,7 +41,7 @@ from .exceptions import PublishFailed
 from .logging import get_logger
 from .rpc import rpc_handler
 from .source_metric import ChunkSize, SourceMetric
-from .types import Metric, Timestamp
+from .timeseries import Metric, Timestamp
 
 logger = get_logger(__name__)
