+++ metricq/synchronous_source.py
@@ -33,7 +33,7 @@ from typing import Any, Dict
 
 from .logging import get_logger
 from .source import MetadataDict, Source
-from .types import Timestamp
+from .timeseries import Timestamp
 
 logger = get_logger(__name__)
