+++ metricq/timeseries/timedelta.py
@@ -1,44 +1,10 @@
-# Copyright (c) 2018, ZIH,
-# Technische Universitaet Dresden,
-# Federal Republic of Germany
-#
-# All rights reserved.
-#
-# Redistribution and use in source and binary forms, with or without modification,
-# are permitted provided that the following conditions are met:
-#
-#     * Redistributions of source code must retain the above copyright notice,
-#       this list of conditions and the following disclaimer.
-#     * Redistributions in binary form must reproduce the above copyright notice,
-#       this list of conditions and the following disclaimer in the documentation
-#       and/or other materials provided with the distribution.
-#     * Neither the name of metricq nor the names of its contributors
-#       may be used to endorse or promote products derived from this software
-#       without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
-# A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR
-# CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
-# EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
-# PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
-# PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
-# LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
-# NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
 import datetime
 import re
-from dataclasses import dataclass
 from functools import total_ordering
-from typing import Any, Dict, Iterator, Union, overload
+from typing import TYPE_CHECKING, Union, overload
 
-from dateutil import tz
-from dateutil.parser import isoparse as dateutil_isoparse
-
-from . import history_pb2
-from .exceptions import NonMonotonicTimestamps
+if TYPE_CHECKING:
+    from .timestamp import Timestamp
 
 
 @total_ordering
@@ -361,405 +327,3 @@ class Timedelta:
         if isinstance(other, datetime.timedelta):
             return self.timedelta < other
         return self._value < other._value
-
-
-@total_ordering
-class Timestamp:
-    """A MetricQ Timestamp
-
-    Args:
-        nanoseconds: number of nanoseconds elapsed since the UNIX epoch
-    """
-
-    _EPOCH = datetime.datetime(1970, 1, 1, tzinfo=datetime.timezone.utc)
-
-    def __init__(self, nanoseconds: int):
-        self._value = nanoseconds
-        """Timestamp value in nanoseconds since :const:`Timestamp._EPOCH`,
-
-        :meta private:
-        """
-
-    @classmethod
-    def from_posix_seconds(cls, seconds: float) -> "Timestamp":
-        """Create a Timestamp from a POSIX timestamp
-
-        Args:
-            seconds: number of seconds since the UNIX epoch
-
-        Returns:
-            :class:`Timestamp`
-        """
-        return Timestamp(int(seconds * 1e9))
-
-    @classmethod
-    def from_datetime(cls, dt: datetime.datetime) -> "Timestamp":
-        """
-        Create a Timestamp from an aware datetime object. If you have a naive
-        datetime object, consider using :meth:`from_local_datetime` instead.
-        """
-        if dt.tzinfo is None:
-            raise TypeError(
-                "cannot to parse naive datetime with `from_datetime`, "
-                "consider using `from_local_datetime` if applicable"
-            )
-        delta = dt - Timestamp._EPOCH
-        seconds = (delta.days * 24 * 3600) + delta.seconds
-        microseconds = seconds * 1000000 + delta.microseconds
-        return Timestamp(microseconds * 1000)
-
-    @classmethod
-    def from_local_datetime(cls, dt: datetime.datetime) -> "Timestamp":
-        """
-        Create a timestamp from a naive :class:`datetime` object that uses the
-        local timezone.
-        If you have a proper aware object, use :meth:`from_datetime` instead.
-
-        This function uses `dateutil.tz.gettz()` which uses `/etc/localtime`.
-        Using naive datetime object can be errorprone. Naive local times are
-        ambiguous during daylight savings time adjustments.
-        If you have to use this function, your workflow is probably broken.
-        """
-        if dt.tzinfo is not None:
-            raise TypeError(
-                "refusing to parse aware datetime with `from_local_datetime`, "
-                "use `from_datetime` instead."
-            )
-        return cls.from_datetime(dt.replace(tzinfo=tz.gettz()))
-
-    @classmethod
-    def from_iso8601(cls, iso_string: str) -> "Timestamp":
-        """Create a Timestamp from a `ISO 8601` date-time string.
-
-        >>> Timestamp.from_iso8601("1970-01-01T00:00:00.0Z") == Timestamp(0)
-        True
-
-        This is a convenience method that parses the date-time string into a
-        :class:`python:datetime.datetime` using
-        :meth:`dateutil:dateutil.parser.isoparse`,
-        and then calls :meth:`from_datetime` to create a :class:`Timestamp` from that.
-
-        The provided `iso_string` must include timezone information. To parse
-        local time strings, you must convert them yourself and use
-        :meth:`from_local_datetime`. Or better yet, somehow create an aware
-        :class:`python:datetime.datetime` and use :meth:`from_datetime`.
-
-        Note:
-            The parser only supports up to *6 sub-second digits*,
-            further digits are simply dropped.
-            If you need to parse timestamps with higher precision,
-            you need to convert the string to nanoseconds yourself.
-
-        Args:
-            iso_string: a date-time string in `ISO 8601` format including a
-                        timezone specifier.
-        """
-        dt = dateutil_isoparse(iso_string)
-        if dt.tzinfo is None:
-            raise TypeError("provided timestamp does not include timezone info")
-        return cls.from_datetime(dt)
-
-    @classmethod
-    def ago(cls, delta: Timedelta) -> "Timestamp":
-        """Return a timestamp `delta` in the past.
-
-        This is equivalent to::
-
-            Timestamp.now() - delta
-
-        Args:
-            delta: the time difference to now
-        """
-        return cls.now() - delta
-
-    @classmethod
-    def from_now(cls, delta: Timedelta) -> "Timestamp":
-        """Return a timestamp `delta` in the future.
-
-        This is equivalent to::
-
-            Timestamp.now() + delta
-
-        Args:
-            delta: the time difference to now
-        """
-        return cls.now() + delta
-
-    @classmethod
-    def now(cls) -> "Timestamp":
-        """Return a Timestamp corresponding to "now"
-
-        Returns:
-            :class:`Timestamp`
-        """
-        return cls.from_datetime(datetime.datetime.now(datetime.timezone.utc))
-
-    @property
-    def posix_ns(self) -> int:
-        """Number of nanoseconds since the UNIX epoch"""
-        return self._value
-
-    @property
-    def posix_us(self) -> float:
-        """Number of microseconds since the UNIX epoch"""
-        return self._value / 1000
-
-    @property
-    def posix_ms(self) -> float:
-        """Number of milliseconds since the UNIX epoch"""
-        return self._value / 1000000
-
-    @property
-    def posix(self) -> float:
-        """Number of seconds since the UNIX epoch"""
-        return self._value / 1000000000
-
-    @property
-    def datetime(self) -> datetime.datetime:
-        """Create an aware UTC datetime object
-
-        All MetricQ timestamps are POSIX timestamps, hence UTC.
-        """
-        # We use timedelta in the hope that this doesn't break
-        # on non-POSIX systems, where fromtimestamp apparently may omit leap seconds
-        # but our MetricQ timestamps are true UNIX timestamps without leap seconds
-        microseconds = self._value // 1000
-        return Timestamp._EPOCH + datetime.timedelta(microseconds=microseconds)
-
-    def __add__(self, delta: Timedelta) -> "Timestamp":
-        """Return the timestamp `delta` in the future (or in the past, if `delta` is negative), relative to a :class:`Timestamp` instance.
-
-        >>> epoch = Timestamp(0)
-        >>> a_week_later = epoch + Timedelta.from_string("7 days")
-        >>> str(a_week_later)
-        '[604800000000000] 1970-01-08 01:00:00+01:00'
-
-        Args:
-            delta: a time duration, possibly negative
-
-        Returns:
-            :class:`Timestamp`
-        """
-        return Timestamp(self._value + delta.ns)
-
-    @overload
-    def __sub__(self, other: "Timedelta") -> "Timestamp":
-        ...
-
-    @overload
-    def __sub__(self, other: "Timestamp") -> "Timedelta":
-        ...
-
-    def __sub__(
-        self, other: Union["Timedelta", "Timestamp"]
-    ) -> Union["Timestamp", "Timedelta"]:
-        if isinstance(other, Timedelta):
-            return Timestamp(self._value - other.ns)
-        if isinstance(other, Timestamp):
-            return Timedelta(self._value - other._value)
-        raise TypeError(
-            "Invalid type to subtract from Timestamp: {}".format(type(other))
-        )
-
-    def __lt__(self, other: "Timestamp") -> bool:
-        """Compare whether this timestamp describes a time before another timestamp.
-
-        >>> now = Timestamp.now()
-        >>> later = now + Timedelta.from_s(10)
-        >>> now < later
-        True
-
-        Together with :meth:`__eq__`, all relational operations (``<=``, ``>``, ``!=``, etc.) are supported.
-        Timestamps are `totally ordered` (as supplied by :func:`python:functools.total_ordering`).
-
-        Args:
-            other: another timestamp
-        """
-
-        return self._value < other._value
-
-    def __eq__(self, other: object) -> bool:
-        """Check whether two :class:`Timestamps<Timestamp>` refer to the same instance of time:
-
-        >>> now = Timestamp.now()
-        >>> later = now + Timedelta.from_s(10)
-        >>> now == later
-        False
-
-        Args:
-            other: another timestamp
-        """
-        if not isinstance(other, Timestamp):
-            return NotImplemented
-
-        return self._value == other._value
-
-    def __str__(self) -> str:
-        """Yield a human-readable date-time string in the local timezone:
-
-        >>> # in UTC+01:00, it was already 1 in the night when the UNIX epoch happened
-        >>> str(Timestamp(0))
-        '[0] 1970-01-01 01:00:00+01:00'
-        """
-
-        # Note we convert to local timezone with astimezone for printing
-        return "[{}] {}".format(self.posix_ns, str(self.datetime.astimezone()))
-
-    def __repr__(self) -> str:
-        return f"Timestamp({self.posix_ns})"
-
-
-@dataclass(frozen=True)
-class TimeValue:
-    """A `timestamp-value` pair.
-
-    Unpack it like so:
-
-    >>> tv = TimeValue(Timestamp.now(), 42.0)
-    >>> (timestamp, value) = tv
-
-    """
-
-    __slots__ = ("timestamp", "value")
-
-    timestamp: Timestamp
-    value: float
-
-    def __iter__(self) -> Iterator[Union["Timestamp", float]]:
-        return iter([self.timestamp, self.value])
-
-    def dict(self) -> dict[str, Any]:
-        """
-        returns a dict representing the TimeValue instance.
-        Keys are `timestamp` and `value`
-        """
-        return {
-            "timestamp": self.timestamp.posix_ns,
-            "value": self.value,
-        }
-
-
-@dataclass(frozen=True)
-class TimeAggregate:
-    """Summary of a metric's values within a certain period of time."""
-
-    __slots__ = (
-        "timestamp",
-        "minimum",
-        "maximum",
-        "sum",
-        "count",
-        "integral_ns",
-        "active_time",
-    )
-
-    timestamp: Timestamp
-    """starting time of this aggregate"""
-    minimum: float
-    """minimum value"""
-    maximum: float
-    """maximum value"""
-    sum: float
-    """sum of all values"""
-    count: int
-    """total number of values"""
-    integral_ns: float
-    """Integral of values in this aggregate over its active time, nanoseconds-based"""
-    active_time: Timedelta
-    """time spanned by this aggregate"""
-
-    @staticmethod
-    def from_proto(
-        timestamp: Timestamp, proto: history_pb2.HistoryResponse.Aggregate
-    ) -> "TimeAggregate":
-        return TimeAggregate(
-            timestamp=timestamp,
-            minimum=proto.minimum,
-            maximum=proto.maximum,
-            sum=proto.sum,
-            count=proto.count,
-            integral_ns=proto.integral,
-            active_time=Timedelta(proto.active_time),
-        )
-
-    @staticmethod
-    def from_value(timestamp: Timestamp, value: float) -> "TimeAggregate":
-        return TimeAggregate(
-            timestamp=timestamp,
-            minimum=value,
-            maximum=value,
-            sum=value,
-            count=1,
-            integral_ns=0,
-            active_time=Timedelta(0),
-        )
-
-    @staticmethod
-    def from_value_pair(
-        timestamp_before: Timestamp, timestamp: Timestamp, value: float
-    ) -> "TimeAggregate":
-        """Create a TimeAggregate from a pair of timestamps (class:`Timestamp`) and one value
-
-        Raises:
-            NonMonotonicTimestamps: if the two timestamps are not strictly monotonic
-        """
-        # Can't use (https://github.com/python/mypy/issues/4610)
-        # if timestamp_before >= timestamp:
-        if not timestamp_before < timestamp:
-            raise NonMonotonicTimestamps(
-                "Timestamps in HistoryResponse are not strictly monotonic ({} -> {})".format(
-                    timestamp_before, timestamp
-                )
-            )
-        delta = timestamp - timestamp_before
-        return TimeAggregate(
-            timestamp=timestamp_before,
-            minimum=value,
-            maximum=value,
-            sum=value,
-            count=1,
-            integral_ns=delta.ns * value,
-            active_time=delta,
-        )
-
-    @property
-    def integral_s(self) -> float:
-        """Integral of values in this aggregate over its active time, seconds-based"""
-        return self.integral_ns / 1e9
-
-    @property
-    def mean(self) -> float:
-        if self.active_time.ns > 0:
-            return self.mean_integral
-        else:
-            return self.mean_sum
-
-    @property
-    def mean_integral(self) -> float:
-        return self.integral_ns / self.active_time.ns
-
-    @property
-    def mean_sum(self) -> float:
-        return self.sum / self.count
-
-    def dict(self) -> dict[str, Any]:
-        """
-        returns a dict representing the TimeAggregate instance.
-        Keys are `timestamp`, `minimum`, `mean`, `maximum`, and `count`.
-        """
-        return {
-            "minimum": self.minimum,
-            "mean": self.mean,
-            "maximum": self.maximum,
-            "count": self.count,
-            "timestamp": self.timestamp.posix_ns,
-        }
-
-
-Metric = str
-"""Type alias for strings that represent metric names
-"""
-
-JsonDict = Dict[str, Any]
-"""Type alias for dicts that represent JSON data
-"""
