+++ setup.cfg
@@ -13,11 +13,6 @@ classifiers =
     Programming Language :: Python :: 3
 
 [options]
-packages =
-    metricq
-    metricq_proto
-package_dir =
-    metricq_proto = lib/metricq-protobuf
 python_requires = >=3.10
 
 [options.package_data]
